Remove unneeded TODO items
Cipher is kept in SecretInfo data types as it is related to encryption
expected from the record layer.  Client chooses a cipher for early
data from the resumed session.  Server may choose a different cipher
for the rest of the handshake.

The Context record is structured with a polymorphic RecordLayer but
also provides other services.  Functions needing the RecordLayer type
argument will keep the RecordLayer argument.  The Context record could
be split into smaller subparts so the record-layer implementation only
accesses the fields it needs.  But this introduces many changes.